Reworking of simulation script
- Change URLs to work with Docker Compose.
- Set minimum trip duration to 10 seconds.
- Add waitForBackend to check that backend is ready.
- Add break if there is no batch.
- Change calcBatteryDepletion to deplete slower (needs more work).
- Implement MIN_TRIP_DURATION in simulateBikeUpdates
- Rename runSimulation to startSimulation
- Import readline, stdin, stdout and dotenv.
- Use readline, stdin, stdout in an attempt to add interactive user input and output during the simulation (work in progress).
